Altered txn info to use api_key as search param if provided and account if account is provided.
