+++ routes/dkg/create_n_transfer.js
@@ -163,7 +163,7 @@ router.post("/", async function (req, res) {
     //   return;
     // }
 
-    query = `INSERT INTO txn_header (txn_id, progress, approver, key_id, request, blockchain, app_name, txn_description, txn_data, ual, keywords, state, txn_hash, txn_fee, trac_fee, epochs, receiver) VALUES (UUID(),?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`;
+    query = `INSERT INTO txn_header (txn_id, progress, approver, key_id, request, blockchain, app_name, txn_description, data_id, ual, keywords, state, txn_hash, txn_fee, trac_fee, epochs, receiver) VALUES (UUID(),?,?,?,?,?,?,?,UUID(),?,?,?,?,?,?,?,?)`;
     params = [
       "PENDING",
       data.approver,
@@ -172,7 +172,6 @@ router.post("/", async function (req, res) {
       data.blockchain,
       app[0].app_name,
       txn_description,
-      typeof data.asset === 'string' || data.asset instanceof String ? (data.asset) : (JSON.stringify(data.asset)),
       null,
       keywords,
       null,
@@ -182,6 +181,7 @@ router.post("/", async function (req, res) {
       epochs,
       data.receiver,
     ];
+
     await queryDB
       .getData(query, params, network, blockchain)
       .then((results) => {
@@ -206,10 +206,27 @@ router.post("/", async function (req, res) {
       console.error("Error retrieving data:", error);
     });
 
+    query = `INSERT INTO data_header (data_id, asset_data) VALUES (?,?)`;
+    params = [
+      txn[0].data_id,
+      typeof data.asset === 'string' || data.asset instanceof String ? (data.asset) : (JSON.stringify(data.asset)),
+    ];
+
+    await queryDB
+      .getData(query, params, network, blockchain)
+      .then((results) => {
+        //console.log('Query results:', results);
+        return results;
+        // Use the results in your variable or perform further operations
+      })
+      .catch((error) => {
+        console.error("Error retrieving data:", error);
+      });
+
     res.status(200).json({
       success: true,
-      msg: "Create-n-Transfer transaction queued successfully. Please use the receipt below to check its completion status.",
-      receipt: `${txn[0].txn_id}`,
+      msg: "Create-n-Transfer transaction queued successfully. Please use the txn id below to check its status.",
+      txn_id: `${txn[0].txn_id}`,
     });
   } catch (e) {
     console.log(e);
